placeholder for annotation routes.  Moved asset management routes to asset.py
