+++ ooiservices/app/main/annotation.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python
+'''
+Annotation endpoints.
+
+'''
+__author__ = 'M@Campbell'
+
+from flask import jsonify, request, current_app, url_for
+from ooiservices.app.main import api
+from ooiservices.app import db
+from authentication import auth
+from ooiservices.app.models import Annotation
+
+#Placeholder
