+++ ooiservices/app/main/asset.py
@@ -1,0 +1,48 @@
+#!/usr/bin/env python
+'''
+Asset management endpoints.
+
+'''
+__author__ = 'M@Campbell'
+
+from flask import jsonify, request, current_app, url_for
+from ooiservices.app.main import api
+from ooiservices.app import db
+from authentication import auth
+from ooiservices.app.models import PlatformDeployment, InstrumentDeployment
+from ooiservices.app.models import Organization
+
+@api.route('/organization', methods=['GET'])
+def get_organizations():
+    organizations = [o.serialize() for o in Organization.query.all()]
+    return jsonify(organizations=organizations)
+
+
+@api.route('/platformlocation', methods=['GET'])
+def get_platform_deployment_geojson_single():
+    geo_locations = {}
+    if len(request.args) > 0:
+        if ('reference_designator' in request.args):
+            if len(request.args['reference_designator']) < 100:
+                reference_designator = request.args['reference_designator']
+                geo_locations = PlatformDeployment.query.filter(PlatformDeployment.reference_designator == reference_designator).all()
+    else:
+        geo_locations = PlatformDeployment.query.all()
+    if len(geo_locations) == 0:
+        return '{}', 204
+    return jsonify({ 'geo_locations' : [{'id' : geo_location.id, 'reference_designator' : geo_location.reference_designator, 'geojson' : geo_location.geojson} for geo_location in geo_locations] })
+
+@api.route('/display_name', methods=['GET'])
+def get_display_name():
+    # 'CE01ISSM-SBD17'
+    platform_deployment_filtered = None
+    display_name = ''
+    if len(request.args) > 0:
+        if ('reference_designator' in request.args):
+            if len(request.args['reference_designator']) < 100:
+                reference_designator = request.args['reference_designator']
+                platform_deployment_filtered = PlatformDeployment.query.filter_by(reference_designator=reference_designator).first_or_404()
+                display_name = platform_deployment_filtered.proper_display_name
+    if platform_deployment_filtered is None:
+        return '{}', 204
+    return jsonify({ 'proper_display_name' : display_name })
